Update job.js: Fix refresh bug
Previously it only refreshed if there any cells in representing a run
"in progress" or "scheduled".

But there is 2 cases where this causes incorrect display:
* The user is logged in and presses "Reset job".
* The user is logged out and someone else does "Reset job" from another
  screen.

In both situations does the table no longer refresh (because the job is
completed), but the state has in fact changed.

I considered moving the `if` to around the setTimeout invocation, but
then the second case described above would still happen.