CleanupAction: Refactor logic to use the ping system.
Previously we used the maximum run time (which inject.js uses
client side) to determine when clients timed out. Clients that
had been created more that that time ago and have not had results
saved, are timed out for sure. That worked, but it was slow. In
most cases a test only takes a few seconds at most, and if that
failed or if the client never started in the first place, then we
waited 5 minutes (or whatever the max runtime is) to mark it as
timed out and reset the run for another client to run.

Now that we have the ping system we know the second a client
isn't active anymore and can reset it right away.

We're still limited by the interval at which api?action=cleanup
is called (so its not instantly), but certainly an improvement
from timeout (5 minutes) to ping miss (30 seconds by default).

Later, when we're running from the command line in an event-based
fashion, we could simply reset individual runs when their client
emits a timeout event (as opposed to the situation now where we
periodically query the db to find timed out clients).

The new query in CleanupAction is a lot more efficient and more
effective in resetting runs as soon as possible. It does require
a join to another table, but that should be fine.
I've considered making a subquery instead of a join:
 SELECT runresults.id as id FROM runresults WHERE runresults.status = 1 AND runresults.client_id IN (SELECT id FROM clients WHERE clients.updated < '20120928195419');

instead of:

 SELECT runresults.id as id FROM runresults INNER JOIN clients ON runresults.client_id = clients.id WHERE runresults.status = 1 AND clients.updated < '20120928195419';

Could find any noticeable difference in performance.
MySQL EXPLAIN made be believe the join was the better choice.

Also:
* Fixed bug in UserAction.php where a client sometimes vanished
  from the user page and then came back on refresh (even though
  the client in question never skipped a beat). This happened
  when a user requested the /user/ page the same second the
  remote client was sent the ping.

* Added appropiate indexes to the database.
  If you're running an install from git master, you'll have to
  add these indexes manually:

  CREATE INDEX idx_clients_updated ON clients (updated);

  CREATE INDEX ids_runs_results ON run_useragent (results_id);

  CREATE INDEX idx_runresults_status_client ON runresults (status, client_id);

  CREATE INDEX idx_run_useragent_useragent_status_run ON run_useragent (useragent_id, status, run_id);

  And drop:

  DROP INDEX idx_run_useragent_useragent ON run_useragent;