BrowserInfo: Implement wildcard system.
Sometimes we need to tolerate for inconsistent implementation
details when working with different parties. This is such case.

Opera, for instance, has no patch versions (only major and minor).
Instead they roll with 2 digit minor versions (11.60, 11.62, 12.01 etc.)

This makes it hard to compare against when working with
browserstack, which only exposes the first digit.

So in order to have "browser: .., major: .." find a match between
testswarm and browserstack-api/browsers, we need to allow a
wildcard.

Added example in sample settings.