(BREAKING CHANGE) Settings: Separate UA definition from browserSet composition.
This avoids duplication, and fixes the bug where two sets could have
the same ID with a different definition (which blows up the system as
uaIDs are globally unique and can only be defined once).

It also solves the problem of what to do with old definitions that
don't belong to any set anymore, but should still be defined so
that old runs can be re-run.

The answer is: Simply remove them from the browserSets entry, but keep
it in the userAgents index.

This reverts a few line changes from d974ada which change logic
from array_values( $stuff ) to array_keys( (array)$stuff ), since
conf.browserSets is now back to being a numerical array, once again.