Swarmstate: Only list relevant browsers
If a browser has online clients or pending runs, it is always
relevant (even if it is for jobs from before a browserSet may
have changed to no longer include that ua).

Otherwise we only list it if it is inside a browserSet per the
current configuration in localSettings.

So we're omitting browsers that have no active clients, no pending
runs and are not currently part of any browser set (where jobs may
be submitted for at any time).