Merge "icon-cleanup": Composition, add "small" class, update imgs.
Instead of composing device, browser and os with 3 layers on top
of each other, now changing it to a more clean layout.
If there is a device, that image is used, otherwise the browser
image is used (so Chrome > Chrome, Mobile Safari on iPhone > iPhone, iPad > iPad). If an OS is specified, it is overlaid in
a semi-opaque bubble in the corner.

Icons for desktop browsers with no OS classified were clean as
as-is and remain unchanged. Icons for mobile browsers look a lot
better now that the the "anonymous man" image is no longer shown
in the back. The device image is now up-front instead of cut-off
in the corner.

Trim trailing * from the icon labels like "Opera 12.1*".
The wildcard is important for matching, but confusing in the UI.

Joining the runs and re-runs badges on HomePage (separate by space
instead of line-break, prevents float breaking by not having
taller boxes for some). Also removed redundant word "pending"
from the labels, making them more concise and less repetitive.

HomePage: Swapped label and badge bootstrap css classes, were used
the wrong way around.

RunPage: Display icon for the user instead of the swarm item.
So, when joining the swarm with Chrome 23.0/Mac OS X 10.8.2,
that is what is shown in the icon, just like on UserPage.
The canonical match from the swarm registry (which might not have
a minor version and OS specified) is only used for HomePage and
JobPage.

Replace weird "pagination-centered thumbnail" bootstrap class
abuse with the now-classified "well-swarm-icon" style instead.

Standardise the "small" version of the swarm icons instead of
selecting it with context relations (like .well > .swarm-icon).
Adding class ".swarm-icon-small".

Updated iOS icon to the latest version (iOS 6).

Choose versions of the iOS and MacOSX logos that look better when
scaled down.