Bootstrap: Update to v2.2.1, add Responsive module, remove min.css
The default width is 940px. Especially the large swarm-results
tables are much wider than that. They currently break out of the
page, off to the right. Leaving a lot of wasted space to the left
of the centred page container.

With the responsive module included, the container will be widened
as browser and screen allow the tables to be displayed better.

Also got rid of the minified version. These are not high traffic
portals and for developers it is probably more useful to load
the raw files.

Source:
* https://github.com/twitter/bootstrap/tree/v2.2.1/img
* https://github.com/twitter/bootstrap/blob/v2.2.1/js/bootstrap-dropdown.js