CleanupAction: Prevent data corruption if query times out.
Sometimes the mysql server can be done, disconnected, rebooting,
overloaded, timing out, whatever it may be.
In that case it should not do the second query, as it depends on
the first.

Also, including the row ID in exception message in JobAction.
Some jobs have 100s of run results, knowing which one is corrupt
helps a lot.

A transaction (BEGIN; COMMIT;) could be a better solution in the
future.