Addjob: Don't prepopulate authtoken.
This was introduced in ee20d5e81e by accident.

The token system is good, but ee20d5e81e accidentally did introduced
the pre-filling featuer in master, which was supposed to be reserved
for the projects brnach (in which logging in means the user has
authority. Wheraes in the current master anyone can create an acount).