project: Fix pagination bugs
The "Next" pager was using the first row instead of the last
row so it appeared to shift only 1 row each page, instead of
shifting to the next page.

The "Previous" pager was broken in the same way and in addition
it didn't really reverse the results back because the return
value of array_reverse wasn't being used (unlike some other
php array_ methods, array_reverse doesn't modify by reference
but returns a new array).