build: Fix paths in bootstrap
Follows-up a18093e. Before that commit the bootstrap files were
embedded in the TestSwarm hierchy (js/, css/ and /img/ are also
the dist directories of a Bootstrap release).

Afterwards, the were in the same directory, which doesn't well.
And images were left behind (causing requests to made to
externals/img/glyphicons-halflings.png, which is 404 Not Found).