Remove @author years
These are either from myself, or for John. The ones for John
actually weren't there originally, I added them in d39e8f0a1a,
so I'll take the liberty of removing the date from them.

For js/pretty.js, I restored the original file header as it was
when John last modified the file.
<https://github.com/jquery/testswarm/blob/f559efb4/js/pretty.js>

Ref #289