Page: Include a version query parameter in urls to static files
This allows them to be cached by e.g. CloudFare for 24h while still
effectibely allowing immediate cache invalidation when deploying
a new version. This currently causes issues when changes are made
where newer HTML more-or-less depends on a newer CSS. Currently
the .css is cached for 24h, but the PHP response is immediate.