ProjectPage: Remove mention of "Last updated" from description header
This reflects when fields in the project authentication record
have last changed. It's very normal for this to rarely ever change
and isn't very useful to end-users.

It's still in the database and the API, but no longer on the page.

This removes confusion that this may relate to when the project
was last used or when the last job was run. For example, for
jQuery Core it says "Last updated in 2013", eventhough the last
commit and TestSwarm run was just a few minutes ago.