scripts: Add script to abort old runs that cannot be rerun
For jquery.org, we remove the test suites from the build server
after 30 days. So if a run still has a pending re-run after that
date, it should be aborted automatically instead of sending it out
to browsers again which will do nothing other than awaiting a 30s
timeout on a 404 Not Found page.