inject: Use status=ABORTED instead of total=-1 to indicate timeout
Follows-up 57b101b2687, which removed the idea of total=-1 as indicator
of timeout in favor of status=ABORTED. The timeout in the outer frame
from run.js was updated accordingly, but the hearbeat timeout in inject.js
was left behind.

Due to the database schema enforcing total as 'unsigned int default 0',
the -1 may've gotten casted to 0, which resulted in an "unknown error"
since the submission will have status=Finished but error=0, total=0, pass=0,
fail=0.

This might fix issue #316, if it turns out the unknown failures there
are actually timeouts.