Page: Avoid 1000+ queries for projects dropdown menu
Before, we re-used ProjectsAction, which in turn re-uses JobAction, which
in turn queries several hundred results (for the last job of each
project) to termine the status color of the "Last job" column
on the "/projects" page. However, none of that is used by the dropdown
menu.

Replace it with a single DB query. Still re-used, to avoid code
duplication, but no longer re-using the rest of that action class.